# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Inventario de Redes",
        "description": "Control de creación, actualización, visualización y eliminación de registros de usuarios",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/status": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "( Muestra todos los estados de dispositivos registrados)",
                "operationId": "ba26089b59c876a5fa5c5c75513e0d49",
                "responses": {
                    "200": {
                        "description": "Datos recuperados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Datos recuperados exitosamente"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "anyOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Status1Schema"
                                                    },
                                                    {
                                                        "$ref": "#/components/schemas/Status2Schema"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/status/{id}": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "( Muestra datos de un estado de dispositivo identificada por {id})",
                "operationId": "8a6aa2edf816734b6daa1b25f3f529e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos recuperados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Datos recuperados exitosamente"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "number",
                                                    "example": "operativo"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Statuses"
                ],
                "summary": "( Elimina los datos de un estado identificado por {id})",
                "operationId": "5c95658333774ec5e84cde0a11daa1c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registro eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Datos de estado de dispositivo borrados exitosamente"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "200"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "3"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "en mantenimiento"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/status/register": {
            "post": {
                "tags": [
                    "Statuses"
                ],
                "summary": "( Crea un nuevo estado de un dispositivo)",
                "operationId": "6a502db4fcf51d24c1e4e6b261dcc3dc",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "number",
                                        "example": "en mantenimiento"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estado de dispositivo creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro creado exitosamente"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "3"
                                                },
                                                "name": {
                                                    "type": "number",
                                                    "example": "en mantenimiento"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/status/update": {
            "put": {
                "tags": [
                    "Statuses"
                ],
                "summary": "( Actualiza los datos de un estado de dispositivo identificado por {id})",
                "operationId": "5b013a4d1764239ea168682da4422a34",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "number",
                                        "example": "en reparacion"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Datos de estado de dispositivo actualizados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro actualizado exitosamente"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "3"
                                                },
                                                "name": {
                                                    "type": "number",
                                                    "example": "en reparación"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Status1Schema": {
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "operativo"
                    }
                },
                "type": "object"
            },
            "Status2Schema": {
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "2"
                    },
                    "name": {
                        "type": "number",
                        "example": "no operativo"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}